Refactored out useless mappers.
