+++ libs/shared/modules/src/case/case.utils.ts
@@ -1,7 +1,6 @@
 import { CaseOverviewTotalItems, CaseStatusEnum } from '@dmr.is/shared/dto'
-import { withTryCatch } from '@dmr.is/utils'
 
-import { CaseModel } from '../models'
+import { CaseModel } from './models'
 
 type StatusResMapper = {
   [key: string]: string
@@ -9,29 +8,6 @@ type StatusResMapper = {
 
 interface Result {
   [key: string]: number
-}
-
-export const caseStatusMapper = (status: string) => {
-  return withTryCatch(() => {
-    switch (status) {
-      case CaseStatusEnum.Submitted:
-        return CaseStatusEnum.Submitted
-      case CaseStatusEnum.InProgress:
-        return CaseStatusEnum.InProgress
-      case CaseStatusEnum.InReview:
-        return CaseStatusEnum.InReview
-      case CaseStatusEnum.ReadyForPublishing:
-        return CaseStatusEnum.ReadyForPublishing
-      case CaseStatusEnum.Published:
-        return CaseStatusEnum.Published
-      case CaseStatusEnum.Rejected:
-        return CaseStatusEnum.Rejected
-      case CaseStatusEnum.Unpublished:
-        return CaseStatusEnum.Unpublished
-    }
-
-    throw new Error(`Case status<${status}> not found`)
-  }, `Failed to migrate case status with title: ${status}`)
 }
 
 export const statusResMapper: StatusResMapper = {
